updated styling and calendar functionality
