+++ client/components/App.jsx
@@ -4,7 +4,7 @@ import Register from './Register'
 import Login from './Login'
 import {connect} from 'react-redux'
 import {logout} from '../actions/auth'
-import Calendar from './Calendar'
+import Meal from './AddMeal'
 
 
 class App extends React.Component {
@@ -16,7 +16,7 @@ class App extends React.Component {
   render () {
     return (
       <Switch>
-        <Route path ='/calendar' component={Calendar} />
+        <Route path ='/meal' component={Meal} />
         <Route path='/register' component={Register} />
         <Route path="/login" component={Login} />
         <button name='logout' onClick={this.handleLogout} >Log out</button>
